feat: Create dummy_Data file to add it AsyncStorage and refactor addCardToDeack() api function
dummy_Dta file was used to add add data to AsyncStorage. But it wint be used again until we clear the application's data from the async storage.
Because AsyncStorage keep data for indefinite period of time and without expiry date which means data will always be in storage until we clear it